// src/lib/simbee-schema.ts
export interface components {
    schemas: {
// ...
        ContentItemDto: {
            id: string;
            client_id: string;
            author_id: string;
            external_id: string;
            content_type?: string;
            visibility?: string;
            engagements_count?: number;
            /** Format: date-time */
            published_at?: string;
            /** Format: date-time */
            created_at?: string;
            /** Format: date-time */
            updated_at?: string;
        };
// ...
        EngagementDto: {
            id: string;
            client_id: string;
            user_id: string;
            content_item_id: string;
            signal_type_id?: string;
            reference_id?: string;
            /** Format: date-time */
            created_at?: string;
            /** Format: date-time */
            updated_at?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}